assets: move selinux in its own package
no intended change of behavior or content:
renames and code movement only.
