+++ pkg/assets/selinux/policy/ocp_v4.10.cil
