+++ pkg/assets/selinux/policy/ocp_v4.11.cil
