+++ pkg/assets/selinux/policy/ocp_v4.12.cil
