+++ pkg/assets/selinux/selinuxinstall.service.template
